pyproject.toml
[project]
name = "google-adk-alpha-vantage"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a]>=1.27.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "structlog>=25.5.0",
    "ruff>=0.11.0",
    "uvicorn>=0.42.0",
    "a2a-sdk>=0.3.3",
    "tenacity>=9.1.4",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
filterwarnings = [
    "ignore:\\[EXPERIMENTAL\\]:UserWarning",
]
